move messages out of the src folder
